move shuffle code to new file
